+++ .dotnet/src/Custom/Chat/ChatReasoningEffortLevel.cs
@@ -5,5 +5,5 @@ using System.Diagnostics.CodeAnalysis;
 namespace OpenAI.Chat;
 
 [CodeGenModel("CreateChatCompletionRequestReasoningEffort")]
-public readonly partial struct ChatReasoningEffort
+public readonly partial struct ChatReasoningEffortLevel
 {}